fix: Move map to row below FB feed
This is so the map does not overflow onto the FB feed.

It's not necessarily the best solution as a more responsive iframe would be better as highlighted by @martinctc. However, that is a longer fix. This is a fix that works for now. This change won't impact how it looks on mobile but it does impact how it looks on desktops.